feature(fraudulent-flag):add modal view for flagging a Post as fraudulent on marketplace ui
Add modal view for flagging a posted car AD as fraudulent.
make some adjustment to the css files to make to make them modular so as to aid reusability.
Also make some adjustments to the html files so they can react well to the css files modifications.

On branch ft-my-posted-ads-page-ui-165965597
Changes to be committed:
	modified:   UI/css/admin.css
	modified:   UI/css/form-pages.css
	modified:   UI/css/general.css
	modified:   UI/css/index.css
	modified:   UI/css/marketplace.css
	modified:   UI/html-pages/admin.html
	modified:   UI/html-pages/index.html
	modified:   UI/html-pages/marketplace.html
	modified:   UI/html-pages/my-posted-ads.html
	modified:   UI/html-pages/signin.html
	modified:   UI/html-pages/signup.html

[Finishes #165910088]